Put Auto{Color}{Side}ConeStackCommand and Auto{Color}{Side}Group to corresponding folders/directories for Twenty203
