+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/command/autonomous/blue_away/AutoBlueAwayConeStackCommand.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.twenty403.command.autonomous;
+package org.firstinspires.ftc.twenty403.command.autonomous.blue_away;
 
 import org.firstinspires.ftc.twenty403.command.claw.ClawCloseCommand;
 import org.firstinspires.ftc.twenty403.command.claw.ClawOpenCommand;
@@ -8,13 +8,13 @@ import org.firstinspires.ftc.twenty403.subsystem.LiftSubsystem;
 import com.technototes.library.command.SequentialCommandGroup;
 import com.technototes.path.subsystem.MecanumDrivebaseSubsystem;
 
-public class AutoBlueHomeConeStackCommand extends SequentialCommandGroup {
-    public AutoBlueHomeConeStackCommand(MecanumDrivebaseSubsystem drive, LiftSubsystem lift, ClawSubsystem claw) {
+public class AutoBlueAwayConeStackCommand extends SequentialCommandGroup {
+    public AutoBlueAwayConeStackCommand(MecanumDrivebaseSubsystem drive, LiftSubsystem lift, ClawSubsystem claw) {
         super(
-                /*new TrajectorySequenceCommand(drive, Robot.Trajectories.BLUE_HOME_STACK)
+                /*new TrajectorySequenceCommand(drive, Robot.Trajectories.BLUE_AWAY_STACK)
                 .alongWith(new ConeReadyToIntakeCommand(cone)),*/
                 new ClawCloseCommand(claw),
-                /*new TrajectorySequenceCommand(drive, Robot.Trajectories.BLUE_HIGH_JUNCTION_HOME)
+                /*new TrajectorySequenceCommand(drive, Robot.Trajectories.BLUE_HIGH_JUNCTION_AWAY)
                 .alongWith(new ConeReadyToScoreCommand(cone)),*/
                 new ClawOpenCommand(claw));
     }
